requirements.txt
numpy
matplotlib.pyplot
streamlit